id is long in domain tier

+++ domain-tier/src/main/java/com/outdoor/foodcalc/domain/model/meal/MealType.java
@@ -14,7 +14,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class MealType implements IDomainEntity<MealType> {
 
-    private final int typeId;
+    private final long typeId;
     private String name;
 
     @Override
